% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_queries.R
\name{assembleQuery}
\alias{assembleQuery}
\title{Interrogate italian companies DB}
\usage{
assembleQuery(conn, base, search_parameters)
}
\arguments{
\item{conn}{A connection object representing the database connection.}

\item{base}{sth.}

\item{search_parameters}{query parameters.}
}
\value{
A data.table with the query result
}
\description{
This function connects and queries the SQLite DB
}
