% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_utils.R
\name{call_out_note}
\alias{call_out_note}
\title{Call Out Note - Quarto Data Collector}
\usage{
call_out_note(statement, condition, message_ok, message_fail)
}
\arguments{
\item{statement}{a function call to evalaute}

\item{condition}{the expected value to return TRUE}

\item{message_ok}{string to display if TRUE}

\item{message_fail}{string to display if FALSE}
}
\value{
HTML code to be inserted in Quarto with eval='asis'
}
\description{
A special type of call-out for Quarto that retrieves "OK" or "WARNING" depending on the condition and messages
}
\author{
Alejandro Abraham \href{mailto:alejandro@liqi.it}{alejandro@liqi.it}
}
