% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_utils.R
\name{check_files}
\alias{check_files}
\title{Check Files - Quarto Data Collector}
\usage{
check_files(lista)
}
\arguments{
\item{lista}{a list() of file names in a folder using list.files()}
}
\value{
data.table with the number of rows and names of each file in the list
}
\description{
Verify that all files in the folder (coming from our server) are not NULL
}
\author{
Alejandro Abraham \href{mailto:alejandro@liqi.it}{alejandro@liqi.it}
}
