% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{forecast_plot}
\alias{forecast_plot}
\title{Invoices forecast - Model forecast Plot}
\usage{
forecast_plot(DT, type_plot = "1", title_plot = "Weekly Forecast")
}
\arguments{
\item{DT}{is a data.table created with model_prep()}

\item{type_plot}{1 for "Invoices Sent" and 2 for "Invoices Recieved", otherwise a valid HEX Value}

\item{title_plot}{plot's title}
}
\value{
data.table with DT ready for model fitting with prophet
}
\description{
Plot Forecast from Liqi propritory Prophet model
}
\author{
Alejandro Abraham \href{mailto:alejandro@liqi.it}{alejandro@liqi.it}
}
