% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_balancesheet}
\alias{get_balancesheet}
\title{Get balance sheet}
\usage{
get_balancesheet(cerved_id, date, apikey, tipo = "E")
}
\arguments{
\item{cerved_id}{A string representing the Cerved ID.}

\item{date}{A string representing the date in 'dd-MM-yyyy' format.}

\item{apikey}{A string representing the API key.}

\item{tipo}{A string representing the balance type. Default is "E".}
}
\value{
A string containing the HTTP response.
}
\description{
This function returns a balance sheet given a Cerved ID, a date, an API key, and a balance type.
}
