% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_balancesheet_list}
\alias{get_balancesheet_list}
\title{Get balance sheet list}
\usage{
get_balancesheet_list(cerved_id, apikey)
}
\arguments{
\item{cerved_id}{A string representing the Cerved ID.}

\item{apikey}{A string representing the API key.}
}
\value{
A string containing the HTTP response.
}
\description{
This function returns a list of balance sheets given a Cerved ID and an API key.
}
