% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_queries.R
\name{get_company_table}
\alias{get_company_table}
\title{Get Company Table}
\usage{
get_company_table(conn, download = FALSE, BASE_PATH = "", verbose = TRUE)
}
\arguments{
\item{conn}{A connection object representing the database connection.}

\item{BASE_PATH}{A string representing the base path for the output file. Default is "".}

\item{verbose}{A logical indicating whether to print messages. Default is TRUE.}
}
\value{
Invisible NULL.
}
\description{
This function retrieves from liqi-prod server the companies tables,
}
