% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_entity_detail}
\alias{get_entity_detail}
\title{Get entity detail}
\usage{
get_entity_detail(piva, apikey)
}
\arguments{
\item{piva}{A string representing the PIVA number.}

\item{apikey}{A string representing the API key.}
}
\value{
A string containing the HTTP response.
}
\description{
This function returns entity details given a PIVA number and an API key.
}
