% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_outlook_from_llm}
\alias{get_outlook_from_llm}
\title{Get economic outlook from LLM}
\usage{
get_outlook_from_llm(ateco_desc, apikey)
}
\arguments{
\item{ateco_desc}{A string representing the description of the business sector.}

\item{apikey}{A string representing the API key.}
}
\value{
A string containing the HTTP response.
}
\description{
This function returns an economic outlook for a given business sector, based on a description.
}
