% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_profile}
\alias{get_profile}
\title{Get profile}
\usage{
get_profile(cerved_id, apikey)
}
\arguments{
\item{cerved_id}{A string representing the Cerved ID.}

\item{apikey}{A string representing the API key.}
}
\value{
A string containing the HTTP response.
}
\description{
This function returns a profile given a Cerved ID and an API key.
}
