% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_score_CGS}
\alias{get_score_CGS}
\title{Get score CGS}
\usage{
get_score_CGS(cerved_id, apikey, codice_score = "CGS")
}
\arguments{
\item{cerved_id}{A string representing the Cerved ID.}

\item{apikey}{A string representing the API key.}

\item{codice_score}{A string representing the score code. Default is "CGS".}
}
\value{
A string containing the HTTP response.
}
\description{
This function returns a CGS score given a Cerved ID, an API key, and a score code.
}
