% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.Date}
\alias{is.Date}
\title{Check Date type}
\usage{
is.Date(vec)
}
\arguments{
\item{vec}{A vector or string.}
}
\value{
A Boolean
}
\description{
TRUE is all elements of the vector are Date type, FALSE otherwise.
}
\author{
Alejandro Abraham \href{mailto:alejandro@liqi.it}{alejandro@liqi.it}
}
