% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_format.R
\docType{data}
\name{liqi_output_reactable}
\alias{liqi_output_reactable}
\title{Liqi Output reactable}
\format{
An object of class \code{reactableTheme} of length 3.
}
\usage{
liqi_output_reactable
}
\value{
an object with lists to pass into the reactable function or options.
}
\description{
Contains the parameters and colour for a standarized table theme.
}
\examples{
library(liqirisk)
library(reactable)
reactable(iris, theme = liqi_output_reactable,
         highlight = TRUE, striped = TRUE, filterable = TRUE, resizable = TRUE, wrap = FALSE)

}
\author{
Alejandro Abraham \href{mailto:alejandro@liqi.it}{alejandro@liqi.it}
}
\keyword{datasets}
