% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collector.R
\name{lister_bs}
\alias{lister_bs}
\title{List of available Balance Sheets - Builder - Data Collector}
\usage{
lister_bs(num_bs, lista)
}
\arguments{
\item{num_bs}{...}

\item{lista}{the returned API call}
}
\value{
a data.table containing the balance sheets info by row
}
\description{
Retrieve from a given CERVD_ID (API Call get_balancesheet_list) the Balance Sheets available, type and date.
}
\author{
Alejandro Abraham \href{mailto:alejandro@liqi.it}{alejandro@liqi.it}
}
