% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_queries.R
\name{load_query}
\alias{load_query}
\title{Load SQL Query}
\usage{
load_query(company, table)
}
\arguments{
\item{company}{A string representing the company name.}

\item{table}{A string representing the table name.}
}
\value{
A string representing the SQL query.
}
\description{
This function generates a SQL query based on a company name and a table name.
}
