% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_forecast}
\alias{model_forecast}
\title{Invoices forecast - Model forecast (Prophet)}
\usage{
model_forecast(DT, col_forecast, n_weeks = 26, complete_dt = FALSE)
}
\arguments{
\item{DT}{is a data.table created with model_prep()}

\item{col_forecast}{is column from DT to be indicated as target}

\item{n_weeks}{number of weeks ahead to forecast}

\item{complete_dt}{boolean to return full prophet data.table output or same DT from model_prep() with n_weeks forecast}
}
\value{
data.table with DT ready for model fitting with prophet
}
\description{
Fit Liqi proprietory Prophet Model application
}
\author{
Alejandro Abraham \href{mailto:alejandro@liqi.it}{alejandro@liqi.it}
}
