% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_prep}
\alias{model_prep}
\title{Invoices forecast - Model preparation}
\usage{
model_prep(DT, c_ds = "ds", c_y = "y")
}
\arguments{
\item{DT}{is a data.table containing invoices from "ricevute.csv" or "emesse.csv" probably}

\item{c_ds}{is column from DT to be indicated as date in the prophet model}

\item{c_y}{is column from DT to be indicated as y (target) in the prophet model}
}
\value{
data.table with DT ready for model fitting with prophet
}
\description{
Prepare data.table for model fitting
}
\author{
Alejandro Abraham \href{mailto:alejandro@liqi.it}{alejandro@liqi.it}
}
