% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_queries.R
\name{open_conn}
\alias{open_conn}
\title{SQL Open Connection}
\usage{
open_conn(
  ssh = "alejandro@prod.bastion.liqi.it",
  url =
    "5432:prod-db-primary-20230426144532573700000011.cct8rb1f18rq.eu-south-1.rds.amazonaws.com:5432"
)
}
\arguments{
\item{ssh}{the user}

\item{url}{the complete string to the DB}
}
\value{
Open SQL Connection
}
\description{
This function connects to a SQL DB
}
