% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_queries.R
\name{query_company_data}
\alias{query_company_data}
\title{Query Company Data}
\usage{
query_company_data(company, conn, BASE_PATH = "", verbose = TRUE)
}
\arguments{
\item{company}{A string representing the company name.}

\item{conn}{A connection object representing the database connection.}

\item{BASE_PATH}{A string representing the base path for the output files. Default is "".}

\item{verbose}{A logical indicating whether to print messages. Default is TRUE.}
}
\value{
Invisible NULL.
}
\description{
This function executes a series of SQL queries based on a company name,
writes the results to CSV files, and optionally prints messages.
}
