% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_client_latest}
\alias{s3_client_latest}
\title{AWS S3 Client lastest date}
\usage{
s3_client_latest(company, risk_bucket = "s3://risk-clientele/clients")
}
\arguments{
\item{company}{the company's LID}

\item{risk_bucket}{the aws bucket}
}
\value{
directory imported from S3
}
\description{
Get the client's  directory from the S3 Risk bucket
}
\author{
Alejandro Abraham \href{mailto:alejandro@liqi.it}{alejandro@liqi.it}
}
