% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_retrieve_externals}
\alias{s3_retrieve_externals}
\title{AWS S3 Bucket Download client external files}
\usage{
s3_retrieve_externals(
  company,
  date,
  risk_bucket = "s3://risk-clientele/clients"
)
}
\arguments{
\item{company}{the company's LID}

\item{date}{the date desired to download (or 'latest')}

\item{risk_bucket}{the aws bucket}
}
\value{
directory imported from S3
}
\description{
Get the client's files from external sources in the S3 Risk bucket
}
\author{
Alejandro Abraham \href{mailto:alejandro@liqi.it}{alejandro@liqi.it}
}
