% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_retrieve_object}
\alias{s3_retrieve_object}
\title{AWS S3 Bucket Download client's file}
\usage{
s3_retrieve_object(
  object_path,
  company_path,
  risk_bucket = "s3://risk-clientele/clients"
)
}
\arguments{
\item{object_path}{the object's full path in the S3 bucket}

\item{company_path}{the bucket's S3 internal path of the company}

\item{risk_bucket}{the aws bucket}
}
\value{
file imported from S3
}
\description{
Get a file from the client's directory in the S3 Risk bucket
}
\author{
Alejandro Abraham \href{mailto:alejandro@liqi.it}{alejandro@liqi.it}
}
