% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_list.R
\docType{data}
\name{table_list}
\alias{table_list}
\title{Table List}
\format{
A data table with 2 variables:
\describe{
\item{table}{character}
\item{name}{character}
}
}
\source{
\url{your_source}
}
\usage{
table_list
}
\description{
A data table containing table names and their corresponding names.
}
\keyword{datasets}
