% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_utils.R
\name{title_note}
\alias{title_note}
\title{Title Note - Quarto Data Collector}
\usage{
title_note(title, hlevel)
}
\arguments{
\item{title}{string containing the Title}

\item{hlevel}{an interger giving the Header level (1-4)}
}
\value{
HTML code to be inserted in Quarto with eval='asis'
}
\description{
A special type of Header for Quarto that can be used when Titles depend on an executed Output from R or Python
}
\author{
Alejandro Abraham \href{mailto:alejandro@liqi.it}{alejandro@liqi.it}
}
