% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_retrieve_directory}
\alias{s3_retrieve_directory}
\title{AWS S3 Bucket Download client directory}
\usage{
s3_retrieve_directory(
  company,
  date,
  dest_path,
  data_type,
  risk_bucket = "s3://risk-clientele/clients",
  zipped = FALSE
)
}
\arguments{
\item{company}{the company's LID}

\item{date}{the date desired to download (or 'latest')}

\item{dest_path}{the local path}

\item{data_type}{internal or external}

\item{risk_bucket}{the aws bucket}

\item{zipped}{create zip folder}
}
\value{
directory imported from S3
}
\description{
Get the client's  directory from the S3 Risk bucket
}
\author{
Alejandro Abraham \href{mailto:alejandro@liqi.it}{alejandro@liqi.it}
}
